[Makefile][macOS] Use pkg-config for libarchive.
In macOS, if libarchive is installed via homebrew,
it does not link well with CMake. The only workaround to
this is to give the absolute path of the libarchive pkg
config (.pc) file to find the actual libarchive installed
via brew. If this is not done, cmake always seems to find
the libarchive installed in system by default.