[Bugfix] Support older qt versions till 5.6.0
Recent PR uses a QString method which is only
supported in Qt 5.10.x and up. Remove that.
Also add CI/CD testing with Qt 5.9.x to ensure
that this does not happen in the future. Can't
find Qt 5.6.x in GH action. Maybe, we can make
a way to install 5.6.x in the future.

fix